📰 🌊 Log multi plataforma (08).
